Refactor test imports to use lib directory structure
