+++ src/tests/globals.test.ts
@@ -1,5 +1,5 @@
-import { sharedWebSockets, resetWebSockets } from "./globals";
-import { WebSocketLike } from "./types";
+import { sharedWebSockets, resetWebSockets } from "../lib/globals";
+import { WebSocketLike } from "../lib/types";
 
 const FIRST_URL = "ws://localhost:1234";
 const SECOND_URL = "ws://localhost:4321";
